requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
pyserial>=3.5
pyyaml>=6.0
pytest>=7.4.0
matplotlib>=3.7.0
pytest-mock>=3.11.0

